refactor assert_schema into utility method for all schema tests to use
