+++ node/src/components/rest_server.rs
@@ -191,46 +191,12 @@ impl Finalize for RestServer {
 
 #[cfg(test)]
 mod schema_tests {
-    use std::{env, fs, io::Write};
-
-    use assert_json_diff::{assert_json_eq, assert_json_matches_no_panic, CompareMode, Config};
-    use schemars::{schema_for, JsonSchema};
-    use serde_json::Value;
-
     use crate::{
         rpcs::{docs::OpenRpcSchema, info::GetValidatorChangesResult},
+        testing::assert_schema,
         types::GetStatusResult,
     };
-
-    fn assert_schema<T: JsonSchema>(schema_path: String) {
-        let expected_schema = fs::read_to_string(&schema_path).unwrap();
-        let expected_schema: Value = serde_json::from_str(&expected_schema).unwrap();
-
-        let actual_schema = schema_for!(T);
-        let actual_schema = serde_json::to_string_pretty(&actual_schema).unwrap();
-        let mut temp_file = tempfile::Builder::new()
-            .suffix(".json")
-            .tempfile_in(env!("OUT_DIR"))
-            .unwrap();
-        temp_file.write_all(actual_schema.as_bytes()).unwrap();
-        let actual_schema: Value = serde_json::from_str(&actual_schema).unwrap();
-        let (_file, path) = temp_file.keep().unwrap();
-        let temp_file_path = path.as_path().clone();
-
-        let result = assert_json_matches_no_panic(
-            &actual_schema,
-            &expected_schema,
-            Config::new(CompareMode::Strict),
-        );
-        assert_eq!(
-            result,
-            Ok(()),
-            "schema does not match:\nexpected:\n{}\nactual:\n{}\n",
-            schema_path,
-            temp_file_path.display()
-        );
-        assert_json_eq!(actual_schema, expected_schema);
-    }
+    use schemars::schema_for;
 
     #[test]
     fn schema_status() {
@@ -238,7 +204,7 @@ mod schema_tests {
             "{}/../resources/test/rest_schema_status.json",
             env!("CARGO_MANIFEST_DIR")
         );
-        assert_schema::<GetStatusResult>(schema_path);
+        assert_schema(schema_path, schema_for!(GetStatusResult));
     }
 
     #[test]
@@ -247,7 +213,7 @@ mod schema_tests {
             "{}/../resources/test/rest_schema_validator_changes.json",
             env!("CARGO_MANIFEST_DIR")
         );
-        assert_schema::<GetValidatorChangesResult>(schema_path);
+        assert_schema(schema_path, schema_for!(GetValidatorChangesResult));
     }
 
     #[test]
@@ -256,6 +222,6 @@ mod schema_tests {
             "{}/../resources/test/rest_schema_rpc_schema.json",
             env!("CARGO_MANIFEST_DIR")
         );
-        assert_schema::<OpenRpcSchema>(schema_path);
+        assert_schema(schema_path, schema_for!(OpenRpcSchema));
     }
 }
