+++ node/src/testing.rs
@@ -13,6 +13,7 @@ mod test_rng;
 use std::{
     any::type_name,
     fmt::Debug,
+    fs,
     marker::PhantomData,
     ops::Range,
     sync::atomic::{AtomicU16, Ordering},
@@ -20,10 +21,12 @@ use std::{
 };
 
 use anyhow::Context;
+use assert_json_diff::{assert_json_eq, assert_json_matches_no_panic, CompareMode, Config};
 use derive_more::From;
 use futures::channel::oneshot;
 use once_cell::sync::Lazy;
 use rand::Rng;
+use serde_json::Value;
 use tempfile::TempDir;
 use tokio::runtime::{self, Runtime};
 use tracing::{debug, warn};
@@ -37,6 +40,8 @@ use crate::{
 };
 pub(crate) use condition_check_reactor::ConditionCheckReactor;
 pub(crate) use multi_stage_test_reactor::MultiStageTestReactor;
+use schemars::schema::RootSchema;
+use std::io::Write;
 pub(crate) use test_rng::TestRng;
 
 /// Time to wait (at most) for a `fatal` to resolve before considering the dropping of a responder a
@@ -357,6 +362,38 @@ pub(crate) fn create_not_expired_deploy(now: Timestamp, test_rng: &mut TestRng) 
         test_rng,
     )
 }
+
+/// Assert that the file at `schema_path` matches the provided `RootSchema`, which can be derived
+/// from `schemars::schema_for!` or `schemars::schema_for_value!`, for example. This method will
+/// create a temporary file with the actual schema and print the location if it fails.
+pub fn assert_schema(schema_path: String, actual_schema: RootSchema) {
+    let expected_schema = fs::read_to_string(&schema_path).unwrap();
+    let expected_schema: Value = serde_json::from_str(&expected_schema).unwrap();
+    let actual_schema = serde_json::to_string_pretty(&actual_schema).unwrap();
+    let mut temp_file = tempfile::Builder::new()
+        .suffix(".json")
+        .tempfile_in(env!("OUT_DIR"))
+        .unwrap();
+    temp_file.write_all(actual_schema.as_bytes()).unwrap();
+    let actual_schema: Value = serde_json::from_str(&actual_schema).unwrap();
+    let (_file, path) = temp_file.keep().unwrap();
+    let temp_file_path = path.as_path().clone();
+
+    let result = assert_json_matches_no_panic(
+        &actual_schema,
+        &expected_schema,
+        Config::new(CompareMode::Strict),
+    );
+    assert_eq!(
+        result,
+        Ok(()),
+        "schema does not match:\nexpected:\n{}\nactual:\n{}\n",
+        schema_path,
+        temp_file_path.display()
+    );
+    assert_json_eq!(actual_schema, expected_schema);
+}
+
 #[cfg(test)]
 mod tests {
     use std::collections::HashSet;
